driver core: Export device_is_bound()
We're about to add the first module user of device_is_bound(), so export
it.  The module user is gpio-max3191x.c, a GPIO driver for Maxim
MAX3191x industrial serializers.  It offers an in-kernel API to change
its operating mode and the API needs to ensure that the driver is bound
to the device and has finished probing before accessing its data
structures.

Cc: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Lukas Wunner <lukas@wunner.de>