serial: pl011: Use threaded interrupt for RX
When not using DMA, the interrupt handler pl011_int() alternatingly
reads 256 characters from the RX FIFO and writes as many characters
as possible from the circular buffer to the TX FIFO.  That's a problem
because:

- If data is constantly sent or received at a high baud rate, interrupts
  will be disabled for prolonged periods, increasing latency for other
  tasks.  In theory the interrupt handler could even get stuck in an
  infinite loop and hog the CPU.

- Because data is received and transmitted alternatingly, the UART's
  full duplex capabilities are not used.

- The situation becomes worse once rs485 support is added because it
  requires that after sending data, the interrupt handler waits for
  the FIFO to empty, waits for the rs485-rts-delay specified in the
  device tree and finally deasserts RTS.  This process can add multiple
  milliseconds to the amount of time spent with interrupts disabled.

Again, if DMA is used, this is mostly a non-issue.  However some PL011
implementations lack DMA, most notably the BCM2835/6/7 (Raspberry Pi).
With 14 million sold as of July 2017, there's a sizeable amount of
affected devices out there.

We can't use request_threaded_irq() here because we need two threads
for RX and TX.  As a first step, introduce a pio_rx kthread which is
woken from the interrupt handler and receives data asynchronously.
It is spawned on ->startup and killed on ->shutdown.

The PL011 does not allow querying the fill level of the RX FIFO, only
whether it is empty.  Previously this was checked before every read from
the RX FIFO, but the present commit adds an optimization:  The RX
interrupt is signaled when the FIFO becomes half full, so the fill level
is known at that point.  The interrupt handler stores this information
in a counter in shared memory and the pio_rx kthread can thus skip the
register access to check emptiness of the FIFO.  Access to the counter
is lockless, the interrupt handler inserts a write memory barrier after
setting it and the pio_rx kthread inserts a data dependency barrier
before reading it.

After reading a character from the FIFO, the pio_rx kthread decrements
the counter.  I believe a memory barrier is not necessary afterwards
because worst case the decrement operation is reordered after a
concurrent store operation by the interrupt handler, merely resulting in
fewer characters that can be read blindly (i.e. without checking FIFO
emptiness).

Once the counter reaches zero, the kthread falls back to checking
emptiness of the FIFO before reading from it.  This is necessary because
characters may arrive as fast as they are read from the FIFO and the RX
interrupt is not signaled earlier than the fill level transitioning past
the "half full" mark.

The interrupt handler previously did not clear RX and RT interrupts
because they were automatically cleared by the hardware once the
interrupt handler emptied the FIFO.  However that happens asynchronously
now, so amend the interrupt handler to clear them.

Cc: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Lukas Wunner <lukas@wunner.de>