serial: pl011: Drop loop from interrupt handler
Previously the interrupt handler alternatingly read bytes from the UART
and wrote bytes to it in a loop until no interrupt would be signaled
anymore.  Now that the interrupt handler just wakes threads, the loop is
no longer necessary as the threads will keep reading or writing from the
FIFOs independently.  Adding debug printk's to the interrupt handler has
shown that the loop is never executed more than once.

Signed-off-by: Lukas Wunner <lukas@wunner.de>