serial: pl011: Document struct uart_amba_port and vendor_data
This is a byproduct of me trying to wrap my head around the driver,
submitted in the hope of making life easier for folks coming after me.

The dmacr element of struct uart_amba_port is unnecessary if
CONFIG_DMA_ENGINE is disabled, therefore moved into the #ifdef section.

I've taken the liberty to shuffle the other elements around slightly
to make the order more logical, at least to my perception.

Signed-off-by: Lukas Wunner <lukas@wunner.de>