net: ks8851: Delay requesting IRQ until opened
The ks8851 driver currently requests the IRQ before registering the
net_device.  Because the net_device name is used as IRQ name and is
still "eth%d" when the IRQ is requested, it's impossibe to tell IRQs
apart if multiple ks8851 chips are present.  Most other drivers delay
requesting the IRQ until the net_device is opened.  Do the same.
The driver doesn't enable interrupts on the chip before the net_device
is opened, so there doesn't seem to be a need to request the IRQ already
on probe.

Cc: Ben Dooks <ben.dooks@codethink.co.uk>
Cc: Tristram Ha <tristram.ha@micrel.com>
Cc: David J. Choi <david.choi@micrel.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>