net: ks8851: Fix register macros
The P1SCLMD_* register macros are misnamed, they pertain to the P1CR
register.  The P1CR_* macros in turn pertain to the P1SR register,
see pages 65 to 68 of the spec:
http://www.hqchip.com/uploads/pdf/201703/47c98946d6c97a4766e14db3f24955f2.pdf

The misnomers have no negative consequences so far because the macros
aren't used by ks8851.c, but that's about to change.

Cc: Ben Dooks <ben.dooks@codethink.co.uk>
Cc: Tristram Ha <tristram.ha@micrel.com>
Cc: David J. Choi <david.choi@micrel.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>