net: ks8851: Support MDI-X control via ethtool
The KSZ8851 supports querying the current MDI-X status through the P1SR
register as well as controlling MDI-X through the P1CR register (or
alternatively P1MBCR).  Afford user space access via ethtool.

Caution: Rev. 2.1 of the KSZ8851SNL specification (published 2013-08-02)
inverted the meaning of the P1CR_FORCEMDIX bit, see pages 3 and 66:
http://www.hqchip.com/uploads/pdf/201703/47c98946d6c97a4766e14db3f24955f2.pdf

Micrel was acquired by Microchip in 2015, which rebranded the spec and
republished it 2017-03-27.  That new spec is based on a pre-2.1 spec and
reintroduces the incorrect meaning of the bit, see page 68:
http://ww1.microchip.com/downloads/en/DeviceDoc/00002381A.pdf

The KSZ8851-16MLL spec contains the correct meaning of the bit in all
known revisions.

Cc: Ben Dooks <ben.dooks@codethink.co.uk>
Cc: Tristram Ha <tristram.ha@micrel.com>
Cc: David J. Choi <david.choi@micrel.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>